// web/src/app/generated/api/v1/sparse_bitset_pb.d.ts
/**
 * Copyright 2026 Google LLC
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

// @generated by protoc-gen-es v2.13.0
// @generated from file api/v1/sparse_bitset.proto (package api.v1, edition 2023)
/* eslint-disable */

import type { GenFile, GenMessage } from '@bufbuild/protobuf/codegenv2';
import type { Message } from '@bufbuild/protobuf';

/**
 * Describes the file api/v1/sparse_bitset.proto.
 */
export declare const file_api_v1_sparse_bitset: GenFile;

/**
 * SparseBitset encodes a sparse set of 32-bit blocks.
 *
 * @generated from message api.v1.SparseBitset
 */
export declare type SparseBitset = Message<'api.v1.SparseBitset'> & {
  /**
   * Block indices (item_id / 32) containing at least one set bit.
   *
   * @generated from field: repeated uint32 indices = 1;
   */
  indices: number[];

  /**
   * 32-bit bitmasks corresponding to each block index where the i-th bit represents (item_id % 32).
   *
   * @generated from field: repeated fixed32 masks = 2;
   */
  masks: number[];
};

/**
 * Describes the message api.v1.SparseBitset.
 * Use `create(SparseBitsetSchema)` to create a new message.
 */
export declare const SparseBitsetSchema: GenMessage<SparseBitset>;